# backend/docs/swagger.yaml
basePath: /
definitions:
  domain.ClipStyle:
    properties:
      background_music_url:
        type: string
      background_music_volume:
        type: number
      brand_logo_position:
        type: string
      brand_logo_scale:
        type: number
      brand_logo_url:
        type: string
      brand_watermark_opacity:
        type: number
      caption_animation:
        type: string
      caption_bg_color:
        type: string
      caption_color:
        type: string
      caption_enabled:
        type: boolean
      caption_font:
        type: string
      caption_max_words:
        type: integer
      caption_position:
        type: string
      caption_size:
        type: integer
      clip_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      original_audio_volume:
        type: number
      overlay_template:
        type: string
      transition_effect:
        type: string
      updated_at:
        type: string
    type: object
...
  utils.ErrorDetail:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
  utils.ErrorResponse:
    properties:
      error:
        properties:
          code:
            type: string
          details:
            items:
              $ref: '#/definitions/utils.ErrorDetail'
            type: array
          message:
            type: string
          request_id:
            type: string
          timestamp:
            type: string
        type: object
    type: object
...
host: localhost:8080
info:
  contact:
    name: API Support
    url: https://example.com/support
  description: API for Reelcut — clip creation from long-form video (auth, projects,
    videos, transcriptions, analysis, clips, templates, jobs).
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://example.com/terms
  title: Reelcut API
  version: "1.0"
paths:
...
  /api/v1/clips:
    get:
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Per page
        in: query
        name: per_page
        type: integer
      - description: Filter by video
        in: query
        name: video_id
        type: string
      - description: Filter by status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: List clips
      tags:
      - clips
    post:
      consumes:
      - application/json
      parameters:
      - description: video_id, name, start_time, end_time, aspect_ratio, ...
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a clip
      tags:
      - clips
  /api/v1/clips/{id}:
    delete:
      parameters:
      - description: Clip ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a clip
      tags:
      - clips
    get:
      parameters:
      - description: Clip ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get clip by ID
      tags:
      - clips
    put:
      consumes:
      - application/json
      parameters:
      - description: Clip ID
        in: path
        name: id
        required: true
        type: string
      - description: name, start_time, end_time, aspect_ratio
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a clip
      tags:
      - clips
  /api/v1/clips/{id}/download:
    get:
      parameters:
      - description: Clip ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Download rendered clip
      tags:
      - clips
  /api/v1/clips/{id}/duplicate:
    post:
      parameters:
      - description: Clip ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Duplicate a clip
      tags:
      - clips
  /api/v1/clips/{id}/render:
    post:
      parameters:
      - description: Clip ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "501":
          description: Not Implemented
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Start clip render job
      tags:
      - clips
  /api/v1/clips/{id}/status:
    get:
      parameters:
      - description: Clip ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Get clip render job status
      tags:
      - clips
  /api/v1/clips/{id}/style:
    get:
      parameters:
      - description: Clip ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get clip style
      tags:
      - clips
    put:
      consumes:
      - application/json
      parameters:
      - description: Clip ID
        in: path
        name: id
        required: true
        type: string
      - description: Style config
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.ClipStyle'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update clip style
      tags:
      - clips
  /api/v1/clips/{id}/style/apply-template/{templateId}:
    post:
      parameters:
      - description: Clip ID
        in: path
        name: id
        required: true
        type: string
      - description: Template ID
        in: path
        name: templateId
        required: true
        type: string
      responses:
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "501":
          description: Not Implemented
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Apply template to clip style
      tags:
      - clips
  /api/v1/jobs:
    get:
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Per page
        in: query
        name: per_page
        type: integer
      - description: Filter by status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: List processing jobs
      tags:
      - jobs
  /api/v1/jobs/{id}:
    get:
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get job by ID
      tags:
      - jobs
  /api/v1/jobs/{id}/cancel:
    post:
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Cancel a job
      tags:
      - jobs
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"